requirements.txt
mysql-connector-python
psycopg2-binary==2.9.9
